chore: reformat

+++ lact-schema/src/profiles.rs
@@ -31,14 +31,15 @@ impl ProfileWatcherState {
 
     pub fn remove_process(&mut self, pid: i32) -> Option<ProfileProcessInfo> {
         if let Some(mut info) = self.process_list.shift_remove(&pid)
-            && let Entry::Occupied(mut entry) = self.process_names_map.entry(info.name) {
-                entry.get_mut().remove(&pid);
-                if entry.get().is_empty() {
-                    let (key, _) = entry.remove_entry();
-                    info.name = key;
-                    return Some(info);
-                }
+            && let Entry::Occupied(mut entry) = self.process_names_map.entry(info.name)
+        {
+            entry.get_mut().remove(&pid);
+            if entry.get().is_empty() {
+                let (key, _) = entry.remove_entry();
+                info.name = key;
+                return Some(info);
             }
+        }
         None
     }
 }
